refactor/qg-196: как-то отрефакторен врокфлоу удаления терапевтической задачи

+++ app/src/main/kotlin/pro/qyoga/app/therapist/therapy/therapeutic_tasks/DeleteTherapeuticTaskWorkflow.kt
@@ -5,7 +5,6 @@ import org.springframework.data.jdbc.core.mapping.AggregateReference
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
 import pro.azhidkov.platform.spring.sdj.erpo.hydration.ref
-import pro.azhidkov.platform.spring.sdj.query.QueryBuilder
 import pro.azhidkov.platform.spring.sdj.withSortBy
 import pro.qyoga.core.clients.journals.JournalEntriesRepo
 import pro.qyoga.core.clients.journals.model.JournalEntry
@@ -20,23 +19,28 @@ class DeleteTherapeuticTaskWorkflow(
 ) : (Long) -> Unit {
 
     override fun invoke(taskId: Long) {
-        ensureNoReferencesExists(taskId)
+        val hasReferencesError = hasReferencesTo(taskId)
+        if (hasReferencesError != null) {
+            throw hasReferencesError
+        }
+
         therapeuticTasksRepo.deleteById(taskId)
     }
 
-    private fun ensureNoReferencesExists(taskId: Long) {
-        val referencesQuery: QueryBuilder.() -> Unit = {
-            JournalEntry::therapeuticTask isEqual AggregateReference.to(taskId)
-        }
-        val referencesExists = journalEntriesRepo.exists(referencesQuery)
+    private fun hasReferencesTo(taskId: Long): TherapeuticTaskHasReferences? {
+        val existingReferences =
+            journalEntriesRepo.findSlice(referencesPageRequest, fetch = JournalEntry.Fetch.client) {
+                JournalEntry::therapeuticTask isEqual AggregateReference.to(taskId)
+            }
+        val task = { therapeuticTasksRepo.findByIdOrNull(taskId) }
 
-        if (referencesExists) {
-            val references =
-                journalEntriesRepo.findAll(referencesPageRequest, fetch = listOf(JournalEntry::client), referencesQuery)
-                    .content
-            val task = therapeuticTasksRepo.findByIdOrNull(taskId)
-                ?: error("References $references to not existing task with id=$taskId")
-            throw TherapeuticTaskHasReferences(task.ref(), references)
+        return if (existingReferences.hasContent()) {
+            return TherapeuticTaskHasReferences(
+                task()?.ref() ?: error("References $existingReferences to not existing task with id=$taskId"),
+                existingReferences.content
+            )
+        } else {
+            null
         }
     }
